feat(settings): add Samba configuration support
